backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
bcrypt==4.1.2
python-dotenv==1.0.0
requests==2.32.3
nltk==3.9.1
pillow==10.4.0
scikit-learn==1.5.2
pandas==2.2.2
joblib==1.4.2
